requirements.txt
fastapi
uvicorn[standard]
langflow
python-dotenv
# Add other libraries your flow specifically needs, for example:
# google-generativeai
# langchain-google-genai
# faiss-cpu
# pypdf
